couchbeam_changes (couchbeam v2.0.0)

Types

changes_option/0

-type changes_option() ::
          continuous | longpoll | normal | include_docs |
          {since, integer()} |
          {timeout, integer()} |
          heartbeat |
          {heartbeat, integer()} |
          {filter, string()} |
          {filter, string(), [{string(), string() | integer()}]} |
          conflicts |
          {style, string()} |
          descending.

changes_options/0

-type changes_options() :: [changes_option()].

db/0

-type db() :: #db{server :: server(), name :: binary(), options :: list()}.

server/0

-type server() :: #server{url :: term(), options :: list()}.

Functions

cancel_stream(Ref)

follow(Db)

-spec follow(Db :: db()) -> {ok, StreamRef :: atom()} | {error, term()}.

follow(Db, Options)

-spec follow(Db :: db(), Options :: changes_options()) -> {ok, StreamRef :: atom()} | {error, term()}.

Stream changes to a pid

Db : a db record

Client : pid or callback where to send changes events where events are The pid receive these events:

{change, StartRef, {done, Lastseq::integer()}

Connection terminated or you got all changes

{change, StartRef, Row :: ejson_object()}

Line of change

{error, LastSeq::integer(), Msg::term()}

Got an error, connection is closed when an error happend.

LastSeq is the last sequence of changes.

While the callbac could be like:

       fun({done, LastSeq}) ->
           ok;
       fun({done, LastSeq}) ->
           ok;
       fun({done, LastSeq}) ->
           ok.

>Options :: changes_stream_options() [continuous
     | longpoll
     | normal
     | include_docs
     | {since, integer() | now}
     | {timeout, integer()}
     | heartbeat | {heartbeat, integer()}
     | {filter, string()} | {filter, string(), list({string(), string() | integer()})}
     | {view, string()},
     | {docids, list))},
     | {stream_to, pid()},
     | {async, once | normal}]

continuous | longpoll | normal: set the type of changes feed to get

include_doc: if you want to include the doc in the line of change

{timeout, Timeout::integer()}: timeout

heartbeat | {heartbeat, Heartbeat::integer()}: set couchdb to send a heartbeat to maintain connection open

{filter, FilterName} | {filter, FilterName, Args::list({key, value})}: set the filter to use with optional arguments

{view, ViewName}: use a view function as filter. Note that it requires to set filter special value "_view" to enable this feature.

>{stream_to, Pid}: the pid where the changes will be sent, by default the current pid. Used for continuous and longpoll connections

Return {ok, StartRef, ChangesPid} or {error, Error}. Ref can be used to disctint all changes from this pid. ChangesPid is the pid of the changes loop process. Can be used to monitor it or kill it when needed.

follow_once(Db)

-spec follow_once(Db :: db()) -> {ok, LastSeq :: integer(), Changes :: list()} | {error, term()}.

follow_once(Db, Options)

-spec follow_once(Db :: db(), Options :: changes_options()) ->
                     {ok, LastSeq :: integer(), Changes :: list()} | {error, term()}.

fetch all changes at once using a normal or longpoll connections.

Db : a db record

Options :: changes_options() [
     | longpoll
     | normal
     | include_docs
     | {since, integer() | now}
     | {timeout, integer()}
     | heartbeat | {heartbeat, integer()}
     | {filter, string()}
     | {filter, string(), list({string(), string() | integer()})}
     | {docids, list()))},
     | {stream_to, pid()}
     ]

longpoll | normal: set the type of changes feed to get

include_docs: if you want to include the doc in the line of change

{timeout, Timeout::integer()}: timeout

heartbeat | {heartbeat, Heartbeat::integer()}: set couchdb to send a heartbeat to maintain connection open

{filter, FilterName} | {filter, FilterName, Args::list({key, value}): set the filter to use with optional arguments

{view, ViewName}: use a view function as filter. Note that it requires to set filter special value "_view" to enable this feature.

Result: {ok, LastSeq::integer(), Rows::list()} or {error, LastSeq, Error}. LastSeq is the last sequence of changes.

stream_next(Ref)