couchbeam_util (couchbeam v2.0.0)

Functions

binary_env(Key, Default)

dbname(DbName)

deprecated(Old, New, When)

encode_att_name(Name)

encode_docid1(DocId)

encode_docid(DocId)

encode_docid_noop(DocId)

encode_query(QSL)

Encode needed query parameter values for JSON

encode_query_value(K, V)

Encode value in JSON if needed depending on the key

force_param(Key, Value, Options)

Replace a value in a property list

get_app_env(Env, Default)

get_value(Key, Prop)

-spec get_value(Key :: term(), Prop :: [term()]) -> term().

Emulate proplists:get_value/2,3 for property lists but use faster lists:keyfind/3

get_value(Key, Prop, Default)

-spec get_value(Key :: term(), Prop :: [term()], Default :: term()) -> term().

oauth_header(Url, Action, OauthProps)

parse_options(Options)

make view options a list

parse_options(Rest, Acc)

propmerge1(L1, L2)

Update a property list with values of the second. In case the same key is in both lists, the value from the first is kept.

propmerge(F, L1, L2)

Merge two property lists (lists of {Key,Value}). All the Key - Value pairs from both lists are included in the new list. If a key occurs in both dictionaries then Fun is called with the key and both values to return a new value. This a wreapper around dict:merge

proxy_header(UserName, Roles, Secret)

proxy_token(Secret, UserName)

shutdown_sync(Pid)

start_app_deps(App)

-spec start_app_deps(atom()) -> ok.

Start depedent applications of App.

to_atom(V)

to_binary(V)

to_integer(V)

to_list(V)