gen_changes behaviour (couchbeam v2.0.0)

gen_changes CouchDB continuous changes consumer behavior This behaviour allows you to create easily a server that consume Couchdb continuous changes

Functions

behaviour_info(_)

call(Name, Request)

call(Name, Request, Timeout)

cast(Dest, Request)

code_change(OldVersion, State, Extra)

get_seq(Pid)

handle_call(Request, From, State)

handle_cast(Msg, State)

handle_info(Info, State)

init(_)

start_link(Module, Db, Options, InitArgs)

create a gen_changes process as part of a supervision tree. The function should be called, directly or indirectly, by the supervisor. InitArgs::list()) -> term() changesoptions() = [changeoption()] changeoption() = {include_docs, string()} | {filter, string()} | {since, integer()|string()} | {heartbeat, string()|boolean()}

stop(Pid)

terminate(Reason, Gen_changes_state)